Move the source generator used for integration tests to a new project
We now require that source generators must live in a netstandard2.0
project, and our various test projects (both unit tests and integration
tests) don't always target that. To keep things simple, we'll have a
single project that provides some trivial source generators so we can
test loading them in all platforms.
