+++ src/Workspaces/TestSourceGenerator/HelloWorldGenerator.cs
@@ -2,18 +2,16 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#nullable disable
-
 using System.Text;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Roslyn.VisualStudio.IntegrationTests
+namespace Microsoft.CodeAnalysis.TestSourceGenerator
 {
     [Generator]
-    internal sealed class IntegrationTestSourceGenerator : ISourceGenerator
+    public sealed class HelloWorldGenerator : ISourceGenerator
     {
-        public const string GeneratedClassName = nameof(IntegrationTestSourceGenerator) + "Output";
+        public const string GeneratedClassName = "HelloWorld";
 
         public void Initialize(GeneratorInitializationContext context)
         {
